+++ packages/riverpod_analyzer_utils_tests/test/legacy_provider_declaration_test.dart
@@ -7,6 +7,46 @@ import 'analyser_test_utils.dart';
 
 void main() {
   group('LegacyProviderDefinition.parse', () {
+    testSource(
+      'Handles unsupported function expressions',
+      source: '''
+import 'package:riverpod/riverpod.dart';
+
+final provider = Provider<int>((ref) => 0);
+
+Provider<int> create() => Provider((ref) => 0);
+final unknown = create();
+
+class Callbable {
+  Provider<int> call() => Provider((ref) => 0);
+}
+
+abstract class A {
+  static Provider<int> create() => Provider((ref) => 0);
+  static final callable = Callbable();
+}
+
+final unknown2 = A.create();
+final unknown3 = A.callable();
+final unknown4 = (() => 42)();
+
+Provider<int> Function() get getter => () => Provider((ref) => 0);
+
+final unknown5 = getter();
+''',
+      (resolver) async {
+        // Regression test for https://github.com/rrousselGit/riverpod/issues/2313
+        final result = await resolver.resolveRiverpodAnalyssiResult();
+
+        expect(result.legacyProviderDeclarations, hasLength(1));
+        expect(result.generatorProviderDeclarations, isEmpty);
+        expect(
+          result.legacyProviderDeclarations.first.node.toSource(),
+          'provider = Provider<int>((ref) => 0)',
+        );
+      },
+    );
+
     testSource('Does not try to parse generated providers',
         runGenerator: true, source: '''
 import 'package:riverpod/riverpod.dart';